support for VLAN based network isolation
Big change in this commit is that OpenVPN now compiles
from source inside of the Docker build proccess. The
purpose of this is to allow me to build in the vlan
community patch.

With the VLAN patch OpenVPN is able to VLAN tag every
packet based on the user CN. In addition it is able to
successfully route to the correct client in the face
of conflicting MAC addresses. Both of these features
are needed to provide full network isolation

In order to eliminate the active layer 2 bridge devices
from the shared user path (which can themselves be attacked)
I have configure the OpenVPN container to use Traffic Control
to send all packets directly from tap0 to the externally
facing veth. From outside the container a vlan interface
can be added to the veth and attached to the challenge
cluster bridge directly. This vastly reduces the attack
surface by eliminating any shared bridging and, as an added
bonus, allows the configuration of rules through tc.