new model for openvpn challenge interface
This commit changes the openvpn container such that it creates
it's own veth pair (rather than relying on Docker to do so)
and pushes one end to the host network namespace. This allows
Naumachia to have an interface it manages (instead of Docker)
and can be linked to the correct bridge without any intermediaries.