users and their connecitons now register in Redis
Whenever a VPN tunnel comes up it will announce its pressence
by adding its hostname (Docker ID) to the 'vpns' set and
adding a 'vpn:<hostname>' hash table in Redis

Whenever a user connects an entry will be added to the 'cnames'
hash table with a user id, and a new hashtable will be added
for the connection.

If this is the clients first time on the tunnel it will pick
an unused vlan at random for them to use.