seperate content (challenges) from the platform
Up until this point I have included all of my work in this one
repo whether it was content (the challenges) or platform. This
commit adds the beginings of a config.yml for Naumachia, which
is currently just used to render the docker-compose file; placing
all of the correct references to the challenge files.

In this commit I have removed the challenges I was working on
(although I have not committed anything I wouldn't want everyone
to know about) and renamed the mitm challenge to be the example.

Going forward my challenges will be in a seperate repo, which will
be private for the purpose of the upcoming competition